Improve the error handling inserted by the CFI pass
Previously, we would change the indirect target to be the `abort` function. This
had two downsides: It was undefined behavior at least in the case of
`indirectbr` and it made it almost impossible to tell what the original jump
target was post-mortem. Now, we instead branch to a proper function call if a
CFI check fails.