Remove `CallBase` visitor from SFI
The visitor only handled calls to `memcpy`. If the call was an intrinsic call,
it will get dispatched to its own visitor. Otherwise, it will be a call to a
function defined either 1) in SVA, in which case the function itself can perform
the check, or 2) in the kernel, in which case it will be defined in C will be
instrumented anyways.