Relax the address space mask for kernel code
Xen runs at the bottom of the upper canonical half, not the top.
Therefore, we need a less restrictive mask.

Note: This is incomplete, we need to also prevent Xen from jumping to
addresses above 0xffff880000000000 as well as those in the lower
canonical half (maybe we could rely on SMEP?).