Decrement/increment unprotected stack pointer in prologues and epilogues
Our split stack implementation separates security-sensitive and
non-sensitive stack objects onto two different stacks.

Security-sensitive objects, such as call return addresses,
compiler-controlled spill slots, and function arguments go on the
protected stack, which lives in SVA-protected memory and is tracked
using the standard stack pointer register (RSP). The target program
(Xen) is prevented from accessing objects on the protected stack using
dynamic (potentially attacker-controlled) pointers or offsets by means
of SFI checks on such memory accesses.

Non-sensitive objects go on a separate unprotected stack which remains
in ordinary memory under the program's control, permitting them to be
accessed via dynamic pointers and offsets. We reserve the RBX register
and use it to track this unprotected stack.

This commit adds the first piece of the compiler-side split stack
implementation. Since we now have two stacks, both stacks need to be
decremented/incremented in function prologues and epilogues. To keep
things simple for the purposes of this prototype, we make both stacks as
large as the single stack would be in a non-split system. This will
allow us (in future commits) to choose which stack an object lives on by
simply adjusting the way references to it are lowered during frame
lowering.

This implementation wastes a bit of space (approaching 100% overhead in
principle), because all local stack objects (not counting return
addresses and callee-saved registers, space for which is allocated with
pushes and pops instead of decrement/increment) now require space on two
stacks, even though they will only actually live on one. This is not a
significant problem in our Xen prototype, because Xen's stack is already
hard-limited to 8 kB for its own design reasons (and Xen uses
per-physical-CPU stacks); thus we will waste a maximum of 8 kB * # of
CPUs, which is minuscule on modern machines.