Move (most) stack objects to unprotected stack in split-stack config
This (assuming we haven't missed anything) completes the compiler side
of split stack support.

Note that, with this now in place, compiling with -msplit-stack *will*
break the target program if it isn't actively aware of and able to
support a split stack. The target program's initialization code (early
boot assembly in Xen's case) will need to set up a usable region of
memory for the unprotected stack and point RBX to it before calling into
any code compiled with -msplit-stack. Any code that explicitly
manipulates the stack (e.g. Xen's "reset stack and jump" operation, or
some of its boot code that moves/copies the stack) will likewise need to
be aware of the split stack and handle both stacks.

To support incremental porting of target programs, I've put a handy
"#if 1" wrapper around the bit of code that actually changes the memory
references to use RBX instead of RSP. Switching it to "#if 0" will let
you compile a program that still reserves RBX as the unprotected stack
pointer, and correctly creates/destroys stack frames on it in function
prologues and epilogues, but which won't break if RBX doesn't yet point
to a usable stack. (This is similar to what would happen if you compiled
with -msplit-stack prior to this commit.)