Fix incorrect computation of unprotected stack offsets
I had incorrectly assumed that "SPAdj" was the magic number LLVM was
using to reflect the difference between SP- and FP-relative frame
reference offsets. It turns out that number is actually used for a
different purpose (and in fact is always 0 when compiling Xen). After
digging into the reference lowering code a bit more I figured out a
"more correct" way of handling this that actually uses an LLVM function
that is *designed* for when you want to force a reference to be lowered
relative to SP instead of the frame pointer.

There are still some "holes" in this implementation, most notably that
we are not handling stack realignment correctly on the unprotected
stack (which is where it matters, since any local objects that might
actually need realignment in C code will be on the unprotected stack;
there's no need for the protected stack to be aligned any more). This is
because we do not currently have a frame pointer for the unprotected
stack. I'm planning to fix that if/when I get the chance (hopefully
before we submit the ASPLOS paper) but in the meantime, I'm hopeful that
the current implementation will be "good enough" to suffice for Xen's
needs.

(The main thing Xen "needs" to use aligned stack objects for is hardware
structures like XSAVE areas. Under the SVA port, SVA should have taken
over responsibility for all of that; the only remaining use of XSAVE
area structures in Xen is for temporary copies which are ultimately
copied into an SVA-owned structure for use with the hardware.  So long
as that's true, Xen doesn't actually need the alignment on its end; it's
only asking the compiler to align the structure because it inherits its
definition of struct xsave_area from SVA, including the alignment
directive. So unless Xen is still doing raw hardware operations with
alignment-sensitive structures somewhere that we missed, it should be
"OK" to break stack alignment for Xen. It'll still work fine for SVA
code because SVA isn't compiling with the split stack.)