Work around frame size mismatch in emitPrologue() vs. emitEpilogue()
Somewhat mystifyingly, the value of NumBytes LLVM computes to determine
the size of a stack pointer decrement/increment doesn't always match
between emitPrologue() and emitEpilogue(). It's not particularly clear
where the discrepancy arises from (as the computation appears exactly
the same between the two functions), but we've only observed this happen
in functions with frame pointers, so my best guess is that LLVM for
whatever reason doesn't bother to keep this consistent when it knows
that it's just going to use the frame pointer anyway to restore RSP in
the epilogue.

A perhaps heavy-handed, but effective, workaround for this is to just
save the value computed by emitPrologue() somewhere that emitEpilogue()
can find it later - namely, in the MachineFrameInfo object, which seems
like the reasonable place to put such a thing (as it's also where the
variables live that form the basis of the NumBytes computation).

I'd really prefer to find out *why* this is happening, but keeping track
of the exact value (rather than recomputing it) is probably a good thing
anyway, since the algorithmic intent is for the SUB and ADD to exactly
match. I'm actually not sure why LLVM didn't do it this way from the
start, although if I knew the answer to that I'd probably know why the
discrepancy is happening. ;-)